change name maches role
