+++ src/main/java/com/fitanywhere/ad/controller/AdController.java
@@ -26,14 +26,14 @@ import org.springframework.web.multipart.MultipartFile;
 
 import com.fitanywhere.ad.model.AdService;
 import com.fitanywhere.ad.model.AdVO;
-import com.fitanywhere.adCarousel.model.AdCarouselService;
-import com.fitanywhere.adCarousel.model.AdCarouselVO;
-import com.fitanywhere.adDate.model.AdDateService;
-import com.fitanywhere.adDate.model.AdDateVO;
+import com.fitanywhere.adate.model.AdDateService;
+import com.fitanywhere.adate.model.AdDateVO;
+import com.fitanywhere.adcarousel.model.AdCarouselService;
+import com.fitanywhere.adcarousel.model.AdCarouselVO;
 
 @Controller
 @RequestMapping("/ad")
-public class adController {
+public class AdController {
 
 	@Autowired
 	private AdService adSvc;
