+++ src/main/java/com/fitanywhere/adate/model/AdDateRepository.java
@@ -1,6 +1,6 @@
 // https://docs.spring.io/spring-data/jpa/docs/current/reference/html/
 
-package com.fitanywhere.adDate.model;
+package com.fitanywhere.adate.model;
 
 import org.springframework.data.jpa.repository.JpaRepository;
