+++ src/main/java/com/fitanywhere/adate/model/AdDateService.java
@@ -1,4 +1,4 @@
-package com.fitanywhere.adDate.model;
+package com.fitanywhere.adate.model;
 
 import java.util.List;
 import java.util.Optional;
