+++ src/main/java/com/fitanywhere/adate/model/AdDateVO.java
@@ -1,4 +1,4 @@
-package com.fitanywhere.adDate.model;
+package com.fitanywhere.adate.model;
 
 import java.sql.Date;
 import java.util.Objects;
