+++ src/main/java/com/fitanywhere/adcarousel/model/AdCarouselRepository.java
@@ -1,10 +1,10 @@
 // https://docs.spring.io/spring-data/jpa/docs/current/reference/html/
 
-package com.fitanywhere.adCarousel.model;
+package com.fitanywhere.adcarousel.model;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface adCarouselRepository extends JpaRepository<AdCarouselVO, Integer> {
+public interface AdCarouselRepository extends JpaRepository<AdCarouselVO, Integer> {
 
 
 }
