+++ src/main/java/com/fitanywhere/adcarousel/model/AdCarouselVO.java
@@ -1,4 +1,4 @@
-package com.fitanywhere.adCarousel.model;
+package com.fitanywhere.adcarousel.model;
 
 import java.sql.Date;
 import java.sql.Timestamp;
