+++ src/main/java/com/fitanywhere/course/model/CourseVO.java
@@ -17,7 +17,7 @@ import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.LastModifiedDate;
 import org.springframework.format.annotation.DateTimeFormat;
 
-import com.fitanywhere.adCarousel.model.AdCarouselVO;
+import com.fitanywhere.adcarousel.model.AdCarouselVO;
 
 /*
  * 註1: classpath必須有javax.persistence-api-x.x.jar
