+++ src/main/resources/templates/front-end/ad/instructor_advertising.html
